Define organization-wide repository labels
The default repository labels (used on issues and PRs) provided by GitHub are not ideal for the needs of this organization.

Having a standardized set of repository labels in all repositories of the organization will make them easier to maintain. This can be handled automatically with a scheduled CI workflow that syncs the labels of all repositories from a centralized source, allowing for repository-specific labels to be merged in where appropriate. This file is that centralized source for the organization-wide labels.