Switch to using YAML for the labels configuration file
The exclusive use of YAML by GitHub Actions workflows, establishes its use as an inevitability. It is useful to
standardize on one format where possible. Since the `github-label-sync` tool supports YAML in addtion to JSON, it is best
to switch to using YAML for the label configuration as well.

The `lannonbr/issue-label-manager-action` GitHub Actions action previously in use required JSON, so I have switched the
workflow to use `github-label-sync` directly. Since this is in the npm package manager, that change does not add any
significant complexity to the workflow. It also allows all the capabilities of the `github-label-sync` tool to be used.