Run "Sync Labels" workflow on relevant PRs
The "Sync Labels" GitHub Actions workflow requires an access token with write permissions in order to change the
repository's labels. For security reasons, the token only has read permissions when a workflow is triggered by a
`pull_request` event. But it is still useful to trigger a workflow run on PRs that affect relevant files in order to
facilitate the review of that PR. The `github-label-sync` tool has a "dry run" mode, which only simulates the changes
that would be taken in regular mode. Switching to tool into this "dry-run" mode allows the entire workflow to run,
providing a fairly effective demonstration of the effect of any proposed changes.