Add support for multiple shared label configuration files
The "universal" labels should be used in all repositories. There also might be additional labels that are specific to a
single repository, and those are defined in the local configuration file. In addition, there might be labels that are
specific to one class of project, so not not appropriate for the universal file, but also applicable to multiple
repositories. Configuration files for these shared labels can be added to the `workflow-templates/assets/sync-labels`
folder, and then referenced in any of the deployments of the workflow by simply adding the filename to the job matrix.